leak fixed

+++ src/draw.c
@@ -6,7 +6,7 @@
 /*   By: evoronin <evoronin@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/03/21 10:46:58 by evoronin      #+#    #+#                 */
-/*   Updated: 2024/03/29 21:56:35 by yitoh         ########   odam.nl         */
+/*   Updated: 2024/03/31 17:36:40 by yitoh         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,15 +30,16 @@ void	paint_line(t_data *data, t_rays *ray, int x, int line_h)
 		wall_x = data->pos_y + ray->perp_wall_dist * ray->ray_dir_y;
 	else
 		wall_x = data->pos_x + ray->perp_wall_dist * ray->ray_dir_x;
-	step = 1 * TEXHEIGHT / line_h;
+	step = 1 * (float)TEXHEIGHT / (float)line_h;
+	// printf ("step:%f = 512/%d     ", step, line_h);
 	tex_pos = (y - (HEIGHT / 2) + (line_h / 2)) * step;
 	t_x = (int)(wall_x * (double)TEXWIDTH);
+	// printf (" text_pos %f = %d - 350 + (%d/2) * %f\n", tex_pos, y, line_h, step);
 	// t_y is actually not moving -> either step is wrong or text_pos original value is too off
 	while (y < y_max)
 	{
 		t_y = (int)tex_pos & (TEXHEIGHT - 1);
 		tex_pos += step;
-		// printf("position t_x %d, t_y:%d    ", t_x, t_y);
 		col = find_pixel(ray, t_x, t_y);
 		mlx_put_pixel(data->img, x, y, col);
 		y++;
